+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+- feat!: Rename tick parameters of 'Market.simulateGas' and 'Semibook.simulateMarketOrder' to 'maxTick'.
+
 # 2.0.0-12
 
 - fix: lazy load ol keys mapping